refactor: feedDetail 상태를 CommentsView에도 전달하기 위해 redux 로직을 상위 컴포넌트로 이동
