+++ client/src/components/template/FeedDetail/FeedDetail.tsx
@@ -1,23 +1,11 @@
 import Avatar from '@Atoms/Avatar'
 import Button from '@Atoms/Button'
 import Image from '@Atoms/Image'
-import { useEffect } from 'react'
-import { useParams } from 'react-router-dom'
+import { FeedDetail } from '@Types/feed'
+import { FC } from 'react'
 import * as S from './FeedDetail.style'
-import { useAppDispatch, useAppSelector } from '@/redux/store'
-import { getFeedDetailAsync } from '@/redux/actions/FeedDetailAction'
 
-const FeedDetail = () => {
-  const { id } = useParams()
-  const dispatch = useAppDispatch()
-  const { body, name, profileImage, pictures } = useAppSelector(
-    state => state.feedDetail.feedDetail,
-  )
-
-  useEffect(() => {
-    dispatch(getFeedDetailAsync(id as string))
-  }, [dispatch])
-
+const FeedDetail: FC<FeedDetail> = ({ body, name, profileImage, pictures }) => {
   return (
     <S.FeedDetailLayout>
       <S.UserInfoBox>
