+++ client/src/pages/Feed/FeedDetail/FeedDetail.tsx
@@ -1,12 +1,26 @@
-import FeedDetailview from '@Template/FeedDetail'
 import styled from 'styled-components'
+import FeedDetailView from '@Template/FeedDetail'
+import CommentsView from '@Template/Comments'
+import { FC, useEffect } from 'react'
+import { useParams } from 'react-router-dom'
+import { useAppDispatch, useAppSelector } from '@/redux/store'
+import { getFeedDetailAsync } from '@/redux/actions/FeedDetailAction'
 
 const FeedDetailLayout = styled.div``
 
-const FeedDetail = () => {
+const FeedDetail: FC = () => {
+  const { id } = useParams()
+  const dispatch = useAppDispatch()
+  const data = useAppSelector(state => state.feedDetail.feedDetail)
+
+  useEffect(() => {
+    dispatch(getFeedDetailAsync(id as string))
+  }, [dispatch])
+
   return (
     <FeedDetailLayout>
-      <FeedDetailview />
+      <FeedDetailView {...data} />
+      <CommentsView {...data} />
     </FeedDetailLayout>
   )
 }
